Reapply "refactor: Migrate CLI hooks commands tests to modular structure"
This reverts commit b85a0bdc4b7fe8f627c8a2d5e8e329786de46907.
