+++ src/cli/hooks/tests/commands/mod.rs
@@ -1,0 +1,12 @@
+//! Modular test structure for CLI hooks commands
+
+pub mod test_utils;
+
+// Test modules organized by functionality
+pub mod add_operations;
+pub mod command_dispatch;
+pub mod list_operations;
+pub mod plugin_management;
+pub mod remove_operations;
+pub mod scripts_directory;
+pub mod validation;
