+++ src/cli/hooks/tests/commands/test_utils/mod.rs
@@ -1,0 +1,8 @@
+//! Test utilities for CLI hooks commands tests
+
+pub mod fixtures;
+pub mod helpers;
+
+// Re-export all utilities for easy access
+pub use fixtures::*;
+pub use helpers::*;
